Editorlocking extraction and MMreload catching
-extracted the Editorlocking into its own file
-added a reset method for the locking
-applied a catching of possible null pointer in OnSaveDisplay
-changed the check for insalled ModuleManager from private to public
