+++ UbioWeldingLtd/UbioWeldingLtd.csproj
@@ -52,6 +52,7 @@
   <ItemGroup>
     <Compile Include="AdvancedGUITextArea.cs" />
     <Compile Include="AdvancedGUITextField.cs" />
+    <Compile Include="EditorLockManager.cs" />
     <Compile Include="OldWeldingPluginConfig.cs" />
     <Compile Include="Constants.cs" />
     <Compile Include="EditorToolbar.cs" />
